More updates to CI/CD, simple test (#43)
* More attempts to fix the CI

* Few more steps to get new  integration framewoek tests

* More CI/CD work

* Finally got simple test to work locally, various fixes in the images

* More test improvements

* Fix flags for integration, do not deploy istio

* make temp dir

* More integration tests integration

* Add policy for integration tests

* More fixes

* Add missing CRD - Galley is crashing if CRDs are missing

* Merge autoinjector

* Circle ci tests

* Add new framework pre-submit tests, they seem to be passed

* Make paths consistent with istio/istio

* Larger timeout, run simple test in one env

* add tmp dir dep

* Fix remote-docker ci

* Fix another path

* Break down the steps of the test, better UI

* In machine we don't have istioctl

* Use explicit go test command, reduce deps on istio/istio

* More crazy paths and deps

* Add junit report, gubernator

* Bookinfo test.sh still too slow, try presubmit

* Make kind until we update the builder image

* Wrong istioctl

* Adjust kubeconfig

+++ bin/test.sh
@@ -21,6 +21,9 @@ function print_help() {
     exit 1
 }
 
+# Customizable
+INGRESS_NS=${INGRESS_NS:-istio-ingress}
+
 export WAIT_TIMEOUT=${WAIT_TIMEOUT:-5m}
 SKIP_CLEANUP=${SKIP_CLEANUP:-0}
 SKIP_SETUP=${SKIP_SETUP:-0}
@@ -61,15 +64,24 @@ BOOKINFO_DEPLOYMENTS="details-v1 productpage-v1 ratings-v1 reviews-v1 reviews-v2
 if [ "$SKIP_SETUP" -ne 1 ]; then
     kubectl create ns bookinfo || true
     # We use the first namespace with sidecar injection enabled to determine the control plane's namespace.
-    ISTIO_CONTROL=$(kubectl get namespaces -o=jsonpath='{$.items[:1].metadata.labels.istio-env}' -l istio-env)
+    # Fails in KIND.
+    if [ -z "$ISTIO_CONTROL" ]; then
+        ISTIO_CONTROL=$(kubectl get namespaces -o=jsonpath='{$.items[:1].metadata.labels.istio-env}' -l istio-env)
+    fi
     ISTIO_CONTROL=${ISTIO_CONTROL:-istio-control}
 
-    kubectl -n bookinfo delete -f samples/bookinfo/platform/kube/bookinfo.yaml --ignore-not-found
-    kubectl -n bookinfo delete -f samples/bookinfo/networking/destination-rule-all.yaml --ignore-not-found
-    kubectl -n bookinfo delete -f samples/bookinfo/networking/bookinfo-gateway.yaml --ignore-not-found
+    if [ -z "$SKIP_DELETE" ]; then
+        kubectl -n bookinfo delete -f samples/bookinfo/platform/kube/bookinfo.yaml --ignore-not-found
+        kubectl -n bookinfo delete -f samples/bookinfo/networking/destination-rule-all.yaml --ignore-not-found
+        kubectl -n bookinfo delete -f samples/bookinfo/networking/bookinfo-gateway.yaml --ignore-not-found
 
-    kubectl label ns bookinfo istio-env=${ISTIO_CONTROL} --overwrite
+        kubectl label ns bookinfo istio-env=${ISTIO_CONTROL} --overwrite
+    fi
 
+    # Must skip if testing with global inject
+    if [ -z "$SKIP_LABEL" ]; then
+        kubectl label ns bookinfo istio-env=${ISTIO_CONTROL} --overwrite
+    fi
     kubectl -n bookinfo apply -f samples/bookinfo/platform/kube/bookinfo.yaml
     kubectl -n bookinfo apply -f samples/bookinfo/networking/destination-rule-all.yaml
     kubectl -n bookinfo apply -f samples/bookinfo/networking/bookinfo-gateway.yaml
@@ -83,12 +95,12 @@ set +e
 n=1
 while true
 do
-    export INGRESS_HOST=$(kubectl -n istio-ingress get service istio-ingressgateway -o jsonpath='{.status.loadBalancer.ingress[0].ip}')
+    export INGRESS_HOST=$(kubectl -n ${INGRESS_NS} get service istio-ingressgateway -o jsonpath='{.status.loadBalancer.ingress[0].ip}')
     if [ -z $INGRESS_HOST ]; then
-        export INGRESS_HOST=$(kubectl -n istio-ingress get service istio-ingressgateway -o jsonpath='{.spec.clusterIP}')
+        export INGRESS_HOST=$(kubectl -n ${INGRESS_NS} get service istio-ingressgateway -o jsonpath='{.spec.clusterIP}')
     fi
-    export INGRESS_PORT=$(kubectl -n istio-ingress get service istio-ingressgateway -o jsonpath='{.spec.ports[?(@.name=="http2")].port}')
-    export SECURE_INGRESS_PORT=$(kubectl -n istio-ingress get service istio-ingressgateway -o jsonpath='{.spec.ports[?(@.name=="https")].port}')
+    export INGRESS_PORT=$(kubectl -n ${INGRESS_NS} get service istio-ingressgateway -o jsonpath='{.spec.ports[?(@.name=="http2")].port}')
+    export SECURE_INGRESS_PORT=$(kubectl -n ${INGRESS_NS} get service istio-ingressgateway -o jsonpath='{.spec.ports[?(@.name=="https")].port}')
     export GATEWAY_URL=$INGRESS_HOST:$INGRESS_PORT
     RESULT=$(curl -s -o /dev/null -w "%{http_code}" http://${GATEWAY_URL}/productpage)
     if [ $RESULT -eq "200"  ]; then
